configure.ac: Use pkg-config instead of buggy m4 script for detecting zlib
The script check_zlib.m4 is currently responsible for finding the zlib
library and headers.

But it is buggy as it overwrites CPPFLAGS with the LDFLAGS at the end.

Instead of fixing this script we use PKG_CHECK_MODULES, which we
already use for libzmq.

Initial patch by Frédéric-Emmanuel Picca.