m4/ac_path_mariadb.m4: Return the mariadb client version as well
The mariadb client package includes its own version number. Let's return
that as well.

We guard that in an #if block so that future package versions without
that define still create a valid test.

We also upgrade from the obsolete AC_TRY_LINK to AC_LINK_IFELSE.

The mariadb client version is then returned in MARIADBCLIENT_VERSION.