assets/configure.ac: Check for a working sed
We are using sed in cppTango/server/idl/Makefile.am.

Our usage of sed is already windling, as it is removed from the tango-9-lts
branch in cppTango.

But we have a report that the stock sed from MacOSX is not working.

So let's just check that it works.

Fails as expected on MacOSX High Sierra 10.13.6.